backend/LeitMind/pyproject.toml
[tool.poetry]
name = "leitmind"
version = "0.1.0"
description = ""
authors = ["Jilvo <jmillot41@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
python-dotenv = "^1.0.1"
fastapi = "^0.115.6"
pydantic = "^2.10.4"
black = "^24.10.0"
flake8 = "^7.1.1"
isort = "^5.13.2"
mypy = "^1.14.0"
pytest = "^8.3.4"
alembic = "^1.14.0"
sqlalchemy = "^2.0.36"
pre-commit = "^4.0.1"
uvicorn = "^0.34.0"
kink = "^0.8.1"
psycopg2-binary = "^2.9.10"
passlib = "^1.7.4"
pyjwt = "^2.10.1"
python-multipart = "^0.0.20"
spacy = "^3.8.3"
lit = "^18.1.8"
numpy = "^2.2.1"
langchain = "^0.3.14"
langchain-community = "^0.3.14"
llama-cpp-python = "^0.3.5"
gradio = "^5.9.1"
torch = "^2.5.1"
transformers = "^4.47.1"
pandas = "^2.2.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"


[tool.black]
line-length = 150
